LPD-90364 feat: Extract useModelArmorTemplateForm hook
