+++ modules/dxp/apps/ai-hub/ai-hub-web/src/main/resources/META-INF/resources/js/model_armor_template_form/ModelArmorTemplateForm.tsx
@@ -7,132 +7,33 @@ import Button from '@clayui/button';
 import ClayForm from '@clayui/form';
 import ClayLayout from '@clayui/layout';
 import Link from '@clayui/link';
-import {openToast} from '@liferay/object-js-components-web';
-import {useFormik} from 'formik';
-import React, {useCallback, useEffect} from 'react';
+import React from 'react';
 
 import './ModelArmorTemplateForm.scss';
 import Toolbar from '../components/ToolBar';
-import {DEFAULT_MODEL_ARMOR_TEMPLATE} from './constants';
 import DetailsPanel from './DetailsPanel';
 import DetectionsPanel from './DetectionsPanel';
 import ResponsibleAIPanel from './ResponsibleAIPanel';
-import {
-	getModelArmorTemplate,
-	putModelArmorTemplate,
-} from './services/ModelArmorTemplateService';
-import {ModelArmorTemplate} from './types/ModelArmorTemplate';
+import {useModelArmorTemplateForm} from './hooks/useModelArmorTemplateForm';
 
 const FORM_ID = 'modelArmorTemplateForm';
+
+interface IProps {
+	accountEntryExternalReferenceCode: string;
+	backURL: string;
+	externalReferenceCode: string;
+}
 
 export default function ModelArmorTemplateForm({
 	accountEntryExternalReferenceCode,
 	backURL,
 	externalReferenceCode,
-}: {
-	accountEntryExternalReferenceCode: string;
-	backURL: string;
-	externalReferenceCode: string;
-}) {
-	const {
-		errors,
-		handleChange,
-		handleSubmit,
-		isSubmitting,
-		setFieldValue,
-		setValues,
-		values,
-	} = useFormik<ModelArmorTemplate>({
-		initialValues: {
-			...DEFAULT_MODEL_ARMOR_TEMPLATE,
-			r_accountToAIHubModelArmorTemplates_accountEntryERC:
-				accountEntryExternalReferenceCode,
-		},
-		onSubmit: async (formValues) => {
-			try {
-				const response = await putModelArmorTemplate(formValues);
-
-				if (response?.externalReferenceCode) {
-					openToast({
-						message: Liferay.Language.get(
-							'model-armor-template-saved-successfully'
-						),
-						type: 'success',
-					});
-				}
-				else {
-					openToast({
-						message: Liferay.Language.get(
-							'failed-to-save-model-armor-template'
-						),
-						type: 'danger',
-					});
-				}
-			}
-			catch (error) {
-				console.error(error);
-
-				openToast({
-					message: Liferay.Language.get(
-						'an-unexpected-error-occurred'
-					),
-					type: 'danger',
-				});
-			}
-		},
-		validate: (formValues) => {
-			const validationErrors: Partial<
-				Record<keyof ModelArmorTemplate, string>
-			> = {};
-
-			if (!formValues.title) {
-				validationErrors.title = Liferay.Language.get('required');
-			}
-
-			if (!formValues.externalReferenceCode) {
-				validationErrors.externalReferenceCode =
-					Liferay.Language.get('required');
-			}
-
-			return validationErrors;
-		},
-	});
-
-	const setField = useCallback(
-		<K extends keyof ModelArmorTemplate>(
-			field: K,
-			value: ModelArmorTemplate[K]
-		) => {
-			setFieldValue(field, value);
-		},
-		[setFieldValue]
-	);
-
-	useEffect(() => {
-		if (!externalReferenceCode) {
-			return;
-		}
-
-		async function fetchFormData() {
-			try {
-				const template = await getModelArmorTemplate(
-					externalReferenceCode
-				);
-
-				setValues(template);
-			}
-			catch (error) {
-				openToast({
-					message: Liferay.Language.get(
-						'failed-to-load-model-armor-template-data'
-					),
-					type: 'danger',
-				});
-			}
-		}
-
-		fetchFormData();
-	}, [externalReferenceCode, setValues]);
+}: IProps) {
+	const {errors, handleSubmit, isSubmitting, setField, values} =
+		useModelArmorTemplateForm({
+			accountEntryExternalReferenceCode,
+			externalReferenceCode,
+		});
 
 	return (
 		<>
@@ -176,7 +77,6 @@ export default function ModelArmorTemplateForm({
 						<ClayLayout.Col md={12}>
 							<DetailsPanel
 								errors={errors}
-								handleChange={handleChange}
 								setField={setField}
 								values={values}
 							/>
